code/develop/opentelemetry_test/requirements.txt
fastapi
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-requests
uvicorn
